Transported some part of code to other modules
