+++ index.html
@@ -7,8 +7,6 @@
   <link rel="stylesheet" href="leaflet/leaflet.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <title>Кексобукинг</title>
-  <!-- <link rel="stylesheet" href="https://unpkg.com/leaflet@1.7.1/dist/leaflet.css" />
-  <script src="https://unpkg.com/leaflet@1.7.1/dist/leaflet.js"></script> -->
 </head>
 <body>
 
@@ -96,7 +94,7 @@
         </fieldset>
         <fieldset class="ad-form__element ad-form__element--wide">
           <label class="ad-form__label" for="address">Адрес (координаты)</label>
-          <input id="address" name="address" type="text" value="35.6895, 139.6917">
+          <input id="address" name="address" type="text" value="35.6895, 139.6917" readonly>
         </fieldset>
         <fieldset class="ad-form__element">
           <label class="ad-form__label" for="type">Тип жилья</label>
@@ -234,7 +232,6 @@
       <button type="button" class="error__button">Попробовать снова</button>
     </div>
   </template>
-  <!-- <script src="leaflet/leaflet.js.map"></script> -->
   <script src="leaflet/leaflet.js"></script>
   <script src="js/main.js" type="module"></script>
 </body>
